% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.mouse.R
\name{coef.mouse}
\alias{coef.mouse}
\title{Shows WCR coefficient vector for a \code{mouse} class object}
\usage{
\method{coef}{mouse}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{mouse} class to be shown.}
}
\value{
Coefficient vector
}
\description{
Shows within-cluster resampling coefficient vector for a \code{mouse} class object.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

