% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mouse.R
\name{confint.mouse}
\alias{confint.mouse}
\title{Shows WCR confidence intervals for a \code{mouse} class object}
\usage{
\method{confint}{mouse}(x, level = 0.95)
}
\arguments{
\item{object}{An object that has the \code{mouse} class to be shown.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows within-cluster resampling confidence intervals for a \code{mouse} class object, which is a list of model outputs.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

