% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mouse_models.R
\name{confint.mouse_models}
\alias{confint.mouse_models}
\title{Shows WCR confidence interval for a \code{mouse_models} class object}
\usage{
\method{confint}{mouse_models}(x, level = 0.95, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{mouse_model} class to be analyzed for confidence interval.}
}
\value{
Matrix containing lower and upper bounds
}
\description{
Shows within-cluster resampling confidence interval for a \code{mouse_models} class object, which is a list of model outputs.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

