% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.wcr.R
\name{confint.wcr}
\alias{confint.wcr}
\title{Shows WCR confidence intervals for a \code{wcr} class object}
\usage{
\method{confint}{wcr}(x, level = 0.95)
}
\arguments{
\item{object}{An object that has the \code{wcr} class to be shown.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows within-cluster resampling confidence intervals for a \code{wcr} class object, which is a list of model outputs.
}
\examples{

## See examples for WcrData and with.wrc_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

