% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.wcr_models.R
\name{confint.wcr_models}
\alias{confint.wcr_models}
\title{Shows WCR confidence interval for a \code{wcr_models} class object}
\usage{
\method{confint}{wcr_models}(x, level = 0.95, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{wcr_model} class to be analyzed for confidence interval.}
}
\value{
Matrix containing lower and upper bounds
}
\description{
Shows within-cluster resampling confidence interval for a \code{wcr_models} class object, which is a list of model outputs.
}
\examples{

## See examples for WcrData and with.wrc_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

