% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse-package.R
\docType{package}
\name{mouse-package}
\alias{mouse}
\alias{mouse-package}
\title{Within-cluster resampling}
\description{
Implements within-cluster resampling
}
\note{
Acknowledgement:

Shunguang Wang for his support in development.

Developmental repository is on github. Your contributions are appreciated.

\url{https://github.com/kaz-yos/mouse}
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida

Maintainer: Kazuki Yoshida <kazukiyoshida@mail.harvard.edu>
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

