% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseFunctions.R
\name{mousemultiples}
\alias{mousemultiples}
\title{Mice Nest Function}
\usage{
mousemultiples(number)
}
\arguments{
\item{number}{If number is larger than 2 you have a nest.
mousemultiples()}
}
\description{
Determines if you have a mice nest or not based on the number of mice you have
}
