% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseFunctions.R
\name{mouseOrrat}
\alias{mouseOrrat}
\title{Mouse or Rat Function}
\usage{
mouseOrrat(nakedtail = FALSE)
}
\arguments{
\item{nakedtail}{Does your rodent have a naked tail? Defaults to FALSE.
mouseOrrat()}
}
\description{
Determines whether you have a mouse or rat based on the parameter
}
