% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool}
\alias{pool}
\title{One-line description}
\usage{
pool(mouse_models)
}
\arguments{
\item{mouse_models}{mouse_models object to be aggregated.}
}
\value{
An object of class \code{mouse}, which is a list of aggregated results.
}
\description{
Short description
}
\details{
detailed description
}
\examples{

## Load
library(mouse)

}
\author{
Kazuki Yoshida
}
\references{
HOffman EB, SEN PK, Weinberg CR. (2001). Within-cluster resampling. \emph{Biometrika}, \bold{88}, 1121-1134.
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

