% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolWcr.R
\name{PoolWcr}
\alias{PoolWcr}
\title{One-line description}
\usage{
PoolWcr(wcr_models)
}
\arguments{
\item{wcr_models}{wcr_models object to be aggregated.}
}
\value{
An object of class \code{wcr}, which is a list of aggregated results.
}
\description{
Short description
}
\details{
detailed description
}
\examples{

## Load
library(wcr)

}
\author{
Kazuki Yoshida
}
\references{
HOffman EB, SEN PK, Weinberg CR. (2001). Within-cluster resampling. \emph{Biometrika}, \bold{88}, 1121-1134.
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

