% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mouse.R
\name{print.mouse}
\alias{print.mouse}
\title{Shows WCR results for a \code{mouse} class object}
\usage{
\method{print}{mouse}(x, printToggle = TRUE, ...)
}
\arguments{
\item{x}{An object that has the \code{mouse} class to be shown.}

\item{printToggle}{whether to print results. Turn off when capturing the invisibly returned matrix.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
Matrix of coefficients, standard errors, Z test statistics, and p-values.
}
\description{
Shows within-cluster resampling results for a \code{mouse} class object.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

