% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mouse_models.R
\name{print.mouse_models}
\alias{print.mouse_models}
\title{Shows WCR results for a \code{mouse_models} class object}
\usage{
\method{print}{mouse_models}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{mouse_model} class to be analyzed for aggregated results.}
}
\value{
None. Results are printed.
}
\description{
Shows within-cluster resampling results for a \code{mouse_models} class object, which is a list of model outputs.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

