% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wcr.R
\name{print.wcr}
\alias{print.wcr}
\title{Shows WCR results for a \code{wcr} class object}
\usage{
\method{print}{wcr}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{wcr} class to be shown.}
}
\value{
None. Results are printed.
}
\description{
Shows within-cluster resampling results for a \code{wcr} class object, which is a list of model outputs.
}
\examples{

## See examples for WcrData and with.wrc_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

