% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mouse_models.R
\name{summary.mouse_models}
\alias{summary.mouse_models}
\title{Shows all results in a \code{mouse_data} class object}
\usage{
\method{summary}{mouse_models}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{mouse_data} class to be shown.}
}
\value{
None. Results are printed.
}
\description{
Shows all model ouputs in a \code{mouse_data} class object, which is a list of model outputs.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

