% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wcr_models.R
\name{summary.wcr_models}
\alias{summary.wcr_models}
\title{Shows all results in a \code{wcr_data} class object}
\usage{
\method{summary}{wcr_models}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{wcr_data} class to be shown.}
}
\value{
None. Results are printed.
}
\description{
Shows all model ouputs in a \code{wcr_data} class object, which is a list of model outputs.
}
\examples{

## See examples for WcrData and with.wrc_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

