% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.mouse.R
\name{vcov.mouse}
\alias{vcov.mouse}
\title{Shows WCR variance covariance matrix for a \code{mouse} class object}
\usage{
\method{vcov}{mouse}(x, ...)
}
\arguments{
\item{...}{For compatibility with generic. Ignored.}

\item{object}{An object that has the \code{mouse} class to be shown.}
}
\value{
Variance covariance matrix
}
\description{
Shows within-cluster resampling variance covariance matrix for a \code{mouse} class object.
}
\examples{

## See examples for mouse and with.mouse_data

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}, \code{\link{with.mouse_data}}
}

