% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcr-package.R
\docType{package}
\name{wcr-package}
\alias{wcr}
\alias{wcr-package}
\title{Within-cluster resampling}
\description{
Implements within-cluster resampling
}
\note{
Acknowledgement:

Shunguang Wang for his support in development.

Developmental repository is on github. Your contributions are appreciated.

\url{https://github.com/kaz-yos/wcr}
}
\examples{

## See examples for WcrData and with.wrc_data

}
\author{
Kazuki Yoshida

Maintainer: Kazuki Yoshida <kazukiyoshida@mail.harvard.edu>
}
\seealso{
\code{\link{WcrData}}, \code{\link{with.wrc_data}}
}

