% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WcrData.R
\name{WcrData}
\alias{WcrData}
\title{One-line description}
\usage{
WcrData(data, cluster_id, Q)
}
\value{
An object of class \code{wcr_data}, which is a list of resampled datasets, each of which consists of independent data points.
}
\description{
Short description
}
\details{
detailed description
}
\examples{

## Load
library(wcr)

}
\author{
Kazuki Yoshida
}
\references{
HOffman EB, SEN PK, Weinberg CR. (2001). Within-cluster resampling. \emph{Biometrika}, \bold{88}, 1121-1134.
}
\seealso{
\code{\link{print.wcr_data}}
}

