% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.mouse_data.R
\name{with.mouse_data}
\alias{with.mouse_data}
\title{Facilitate analyzing \code{mouse_data} class objects}
\usage{
\method{with}{mouse_data}(data, expr, ...)
}
\arguments{
\item{expr}{Some analysis function invocation. Variable in the resampled dataset can be directly called, thus, no need to fill in the \code{data} argument.}

\item{...}{For compatibility with generic. Ignored.}

\item{x}{Object returned by \code{\link{mouse}} function.}
}
\value{
An object with class \code{mouse_models}, which is a list of model output objects.
}
\description{
\code{with} method for the \code{mouse_data} class objects created by \code{\link{mouse}} function.
}
\examples{

## See examples for ...

}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{mouse}}
}

